+++ mock/docs/mock.1
@@ -392,7 +392,7 @@ Requires rpm >= 6.0.91.  Valid values for \fIMODE\fR:
 .RS
 .TP
 .B best_effort
-Use separate %check when rpmbuild supports \-bk, fall back to to normal order otherwise.
+Use separate %check when rpmbuild supports \-bk, fall back to normal order otherwise.
 .TP
 .B enforce
 Require \-bk support; fail if unavailable.
