+++ mock/py/mock.py
@@ -264,7 +264,7 @@ def command_parse():
                       default=True, help="pass --nocheck to rpmbuild to skip 'make check' tests")
     parser.add_option("--separate-check",
                       choices=['best_effort', 'enforce', 'off'],
-                      default='off', dest="separate_check",
+                      default=None, dest="separate_check",
                       help="Run %%check as a separate rpmbuild phase using -bk.  "
                            "'best_effort' uses it when rpmbuild supports -bk, "
                            "falls back silently otherwise.  'enforce' fails if "
